+++ components/shared/webrender/rendering_context.rs
